backend/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
coverage==7.0.5
cryptography==41.0.1
Django==4.1.4
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
Pillow==9.4.0
psycopg2-binary==2.9.6
pycparser==2.21
python-decouple==3.6
pytz==2022.7
sentry-sdk==1.22.2
sqlparse==0.4.3
typing_extensions==4.6.3
urllib3==1.26.15
